fix: keep secured assets on safe swap route
