+++ .changeset/thorchain-dynamic-secured-assets.md
@@ -3,4 +3,4 @@
 '@vultisig/sdk': minor
 ---
 
-Expose a cached dynamic THORChain secured-asset catalog with an offline fallback, and prefer direct `SECURE+` mints for same-underlying destinations.
+Expose a cached dynamic THORChain secured-asset catalog with an offline fallback for swap destination discovery.
