+++ packages/sdk/src/platforms/react-native/index.ts
@@ -86,9 +86,9 @@ export {
   isCosmosMemoWithinCap,
 } from '@vultisig/core-chain/chains/cosmos/cosmosMemoCap'
 
-// Dynamic THORChain secured-asset discovery and same-underlying SECURE+ routing.
-// These fetch-based/pure helpers are RN-safe and intentionally match the root
-// SDK entrypoint so mobile clients consume the same catalog contract.
+// Dynamic THORChain secured-asset discovery. These fetch-based/pure helpers
+// are RN-safe and intentionally match the root SDK entrypoint so mobile
+// clients consume the same catalog contract.
 export type {
   ThorchainSecuredAsset,
   ThorchainSecuredAssetCatalog,
@@ -103,11 +103,6 @@ export {
   parseThorchainSecuredAssets,
   thorchainSecuredAssetFallback,
 } from '@vultisig/core-chain/chains/cosmos/thor/securedAssets'
-export {
-  getThorchainSecuredAssetMintQuote,
-  isSameUnderlyingThorchainSecuredAsset,
-} from '@vultisig/core-chain/swap/native/securedAssetMint'
-
 // XRP Ledger issued-currency canonicals — pure helpers/tables that are safe on
 // the RN graph and should stay in parity with the root SDK entrypoint.
 export {
